style: Move avatar into own file
